added windows support

+++ library/font/src/Typeface.cpp
@@ -37,10 +37,10 @@ RenderedRune Typeface::render_glyph(const int size, const Rune rune) const
            m_face->glyph->bitmap.buffer,
            m_face->glyph->bitmap.width,
            m_face->glyph->bitmap.rows,
-           static_cast<uint32_t>(m_face->glyph->advance.x >> 6),
-           static_cast<uint32_t>(m_face->glyph->advance.y >> 6),
-           static_cast<uint32_t>(m_face->glyph->bitmap_left),
-           static_cast<uint32_t>(m_face->glyph->bitmap_top),
+           m_face->glyph->advance.x >> 6,
+           m_face->glyph->advance.y >> 6,
+           m_face->glyph->bitmap_left,
+           m_face->glyph->bitmap_top,
    };
 }
